refactor(comp:layoutpro): 调整layoutPro代码结构
