+++ packages/pro/layout/src/util/prefix.ts
@@ -1,0 +1,10 @@
+/**
+ * @license
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://github.com/IDuxFE/idux/blob/main/LICENSE
+ */
+
+export const prefixCls = 'ix'
+export const layoutCls = `${prefixCls}-layout-pro`
+export const layoutSiderTriggerCls = `${prefixCls}-layout-pro-ctrl`
